Определите тип данных для переменных. [#172634 #422827] исправление возраст - byte, бактерии - double

+++ src/main/java/ru/job4j/Variables.java
@@ -6,11 +6,11 @@ public class Variables {
         float length = 78.0332F;
         boolean baby = true;
 
-        short ageOfPerson;
+        byte ageOfPerson;
         char gender;
         float height;
         short countOfBricks;
-        int countOfBacteria;
+        double countOfBacteria;
         double distance;
         boolean permission;
         long sizeOfFile;
